pyproject.toml
[project]
name = "scan-to-pdf"
version = "0.1.0"
description = "CLI to combine images into a searchable PDF with OCR"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "pytesseract>=0.3.10",
    "pypdf>=5.0.0",
    "pillow>=11.0.0",
    "tqdm>=4.65.0",
]

[project.scripts]
scan-to-pdf = "scan_to_pdf.main:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "pytest>=8.0.0",
]

[tool.mypy]
ignore_missing_imports = true
strict = false
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []
